Refactoring ExecutorServiceCancel tasks
They are changed so that they implement the correct base class
according to where they run, address or partition.

Also they are not BlockingTask anymore.
