+++ hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/executorservice/ExecutorServiceCancelOnPartitionMessageTask.java
@@ -18,28 +18,25 @@ package com.hazelcast.client.impl.protocol.task.executorservice;
 
 import com.hazelcast.client.impl.protocol.ClientMessage;
 import com.hazelcast.client.impl.protocol.codec.ExecutorServiceCancelOnPartitionCodec;
+import com.hazelcast.client.impl.protocol.task.AbstractPartitionMessageTask;
 import com.hazelcast.executor.impl.DistributedExecutorService;
 import com.hazelcast.executor.impl.operations.CancellationOperation;
 import com.hazelcast.instance.impl.Node;
 import com.hazelcast.nio.Connection;
-import com.hazelcast.spi.impl.operationservice.InvocationBuilder;
-import com.hazelcast.spi.impl.operationservice.impl.OperationServiceImpl;
+import com.hazelcast.spi.impl.operationservice.Operation;
 
-import java.net.UnknownHostException;
+import java.security.Permission;
 
 public class ExecutorServiceCancelOnPartitionMessageTask
-        extends AbstractExecutorServiceCancelMessageTask<ExecutorServiceCancelOnPartitionCodec.RequestParameters> {
+        extends AbstractPartitionMessageTask<ExecutorServiceCancelOnPartitionCodec.RequestParameters> {
 
     public ExecutorServiceCancelOnPartitionMessageTask(ClientMessage clientMessage, Node node, Connection connection) {
         super(clientMessage, node, connection);
     }
 
     @Override
-    protected InvocationBuilder createInvocationBuilder() throws UnknownHostException {
-        final OperationServiceImpl operationService = nodeEngine.getOperationService();
-        final String serviceName = DistributedExecutorService.SERVICE_NAME;
-        CancellationOperation op = new CancellationOperation(parameters.uuid, parameters.interrupt);
-        return operationService.createInvocationBuilder(serviceName, op, clientMessage.getPartitionId());
+    protected Operation prepareOperation() {
+        return new CancellationOperation(parameters.uuid, parameters.interrupt);
     }
 
     @Override
@@ -47,9 +44,13 @@ public class ExecutorServiceCancelOnPartitionMessageTask
         return ExecutorServiceCancelOnPartitionCodec.decodeRequest(clientMessage);
     }
 
-
     protected ClientMessage encodeResponse(Object response) {
         return ExecutorServiceCancelOnPartitionCodec.encodeResponse((Boolean) response);
+    }
+
+    @Override
+    public String getServiceName() {
+        return DistributedExecutorService.SERVICE_NAME;
     }
 
     @Override
@@ -57,4 +58,18 @@ public class ExecutorServiceCancelOnPartitionMessageTask
         return null;
     }
 
+    @Override
+    public Permission getRequiredPermission() {
+        return null;
+    }
+
+    @Override
+    public String getMethodName() {
+        return "cancel";
+    }
+
+    @Override
+    public Object[] getParameters() {
+        return null;
+    }
 }
